+++ src/main/java/org/zerock/ex01/controller/MealPlanController.java
@@ -6,11 +6,13 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+import org.zerock.ex01.dto.CustomRecipeDTO;
 import org.zerock.ex01.dto.GenerateMealPlanDTO;
 import org.zerock.ex01.dto.MealPlanResulttDTO;
 import org.zerock.ex01.service.MealPlanService;
 
 import java.util.List;
+
 
 @RestController
 //@CrossOrigin(origins = "*", allowedHeaders = "*")
@@ -20,49 +22,50 @@ import java.util.List;
 public class MealPlanController {
     private final MealPlanService mealPlanService;
 
-    @GetMapping("checkMealPlanCount")
+    @GetMapping("/checkMealPlanCount")
     public ResponseEntity<?> checkMealPlanCount(@AuthenticationPrincipal String userId) {//Meal plan 개수 체크하기
-        userId = "jacky4257@gmail.com";
+        log.info("checkMealPlanCount_userId {}", userId);
         Long rsCount = mealPlanService.checkMealPlanCount(userId);
-        return ResponseEntity.ok().body(rsCount < 7 ? rsCount == 0 ? "true" : "only day" : "false");//true면 추가 가능
+        return ResponseEntity.ok().body(rsCount < 7 ? rsCount == 0 ? "true" : "onlyDay" : "false");//true면 추가 가능
     }
 
-    @GetMapping("generateMealPlan")
-    public ResponseEntity<?> generateMealPlan(GenerateMealPlanDTO generateMealPlanDTO) {
-        return ResponseEntity.ok().body(mealPlanService.generateMealPlan(generateMealPlanDTO));
+    @PostMapping("/nser/generateMealPlan")
+    public List<Object> generateMealPlan(GenerateMealPlanDTO generateMealPlanDTO) {
+        log.info("generateMealPlanDTO {}", generateMealPlanDTO);
+        return mealPlanService.generateMealPlan(generateMealPlanDTO);
     }
 
-    @PostMapping("addMealPlan")
+    @PostMapping("/addMealPlan")
     public ResponseEntity<?> addMealPlan(@AuthenticationPrincipal String userId, @RequestBody List<MealPlanResulttDTO> mealPlanRequest) {
-        userId = "jacky4257@gmail.com";
-
-        log.info(mealPlanRequest);
+        log.info("addMealPlan {}", mealPlanRequest);
         mealPlanService.addMealPlan(userId, mealPlanRequest);
         return ResponseEntity.ok().body(HttpStatus.OK);
     }
 
-    @GetMapping("getMealPlan")
+    @GetMapping("/getMealPlan")
     public ResponseEntity<?> getMealPlan(@AuthenticationPrincipal String userId) {//유저의 MealPlan select
-        userId = "jacky4257@gmail.com";
-        return ResponseEntity.ok().body(mealPlanService.getMealPlan(userId));
+        log.info("getMealPlan {}", userId);
+        return ResponseEntity.ok().body(mealPlanService.getMealPlanList(userId));
     }
 
-    @GetMapping("removeMealPlan")
-    public ResponseEntity<?> removeMealPlan(Long planId) {
+    @GetMapping("/removeMealPlan")
+    public ResponseEntity<?> removeMealPlan(@AuthenticationPrincipal String userId, Long planId) {
+        log.info("removeMealPlan {}", userId);
+        MealPlanResulttDTO mealPlanResulttDTO = mealPlanService.getMealPlanWithPlanId(planId);
+        if (!mealPlanResulttDTO.getUserEmail().equals(userId))
+            return new ResponseEntity<>(false, HttpStatus.INTERNAL_SERVER_ERROR);
         mealPlanService.removeMealPlan(planId);
-        return ResponseEntity.ok().body(null);
+        return ResponseEntity.ok().body(HttpStatus.OK);
     }
 
-    @GetMapping("ChangeStatus")
+    @GetMapping("/nser/ChangeStatus")
     public ResponseEntity<?> ChangeStatus(Long planListId) {
-
         mealPlanService.ChangeStatus(planListId);
         return ResponseEntity.ok().body(HttpStatus.OK);
     }
 
     @GetMapping("StatusAllReset")
     public ResponseEntity<?> StatusAllReset(@AuthenticationPrincipal String userId) {
-        userId = "jacky4257@gmail.com";
         mealPlanService.StatusAllReset(userId);
         return ResponseEntity.ok().body(HttpStatus.OK);
     }
